+++ src/main/java/com/example/springcrud/repo/book_repo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface book_repo extends JpaRepository<book, Long> {
+public interface book_repo extends JpaRepository<book, Integer> {
 
     
 }
